code/rag-demo/requirements.txt
sentence-transformers
faiss-cpu
numpy
